+++ src/components/signature.js
@@ -56,6 +56,7 @@ module.exports = function(app) {
       $templateCache.put('formio/components/signature/validate.html',
         '<ng-form>' +
           '<form-builder-option property="validate.required"></form-builder-option>' +
+          '<form-builder-option property="validate.customMessage"></form-builder-option>' +
         '</ng-form>'
       );
     }